jpeg funny

+++ shared/src/net.rs
@@ -6,7 +6,7 @@ where
     R: AsyncRead + Unpin,
 {
     // println!("waiting to read!");
-    let mut len_buf = [0u8; size_of::<usize>()];
+    let mut len_buf = [0u8; size_of::<u64>()];
     reader.read_exact(&mut len_buf).await?;
     println!("got length of payload");
     let len = usize::from_le_bytes(len_buf);
@@ -21,7 +21,7 @@ pub async fn write<W>(writer: &mut W, data: &[u8]) -> std::io::Result<()>
 where
     W: AsyncWrite + Unpin,
 {
-    let len_bytes = (data.len() as usize).to_le_bytes();
+    let len_bytes = (data.len() as u64).to_le_bytes();
     writer.write_all(&len_bytes).await?;
     writer.write_all(data).await?;
     writer.flush().await?;
